+++ HISTORY.md
@@ -24,3 +24,7 @@ This meteor package is a fork of [elevatedevdesign:autoform-nouislider](https://
 - prevent potential jump effect on reactive computations in cases, where documents are edited (update form) and a new 
 slider value has been set
 - fix start value on the slider was set to 0 when editing a document with a given value
+
+### 0.6.0
+- import package dynamically or statically
+- support `disabled` state